型の変更

+++ OLED/oled.go
@@ -9,8 +9,8 @@ import (
 // Display wraps
 type Display struct {
 	device ssd1306.Device
-	XPos   int16
-	YPos   int16
+	XPos   uint16
+	YPos   uint16
 	font   *GT20L16J1Y.Device
 }
 
@@ -18,7 +18,7 @@ func NewDisplay(device ssd1306.Device, font *GT20L16J1Y.Device) Display {
 	return Display{device: device, font: font}
 }
 
-func (d *Display) LcdPrint(x int16, y int16, str string) {
+func (d *Display) LcdPrint(x uint16, y uint16, str string) {
 	d.XPos = x // set position X
 	d.YPos = y // set position Y
 	d.printText(str)
@@ -46,13 +46,13 @@ func (d *Display) printText(str string) {
 }
 
 func (d *Display) printChar(font GT20L16J1Y.Font) {
-	var x, y int16
+	var x, y uint16
 	for y = 0; y < font.FontHeight; y++ {
 		for x = 0; x < font.FontWidth; x++ {
 			if font.FontData[x]&(0x01<<y) != 0x00 {
-				d.device.SetPixel(x+d.XPos, y+d.YPos, color.RGBA{255, 255, 255, 0})
+				d.device.SetPixel(int16(x+d.XPos), int16(y+d.YPos), color.RGBA{255, 255, 255, 0})
 			} else {
-				d.device.SetPixel(x+d.XPos, y+d.YPos, color.RGBA{0, 0, 0, 0})
+				d.device.SetPixel(int16(x+d.XPos), int16(y+d.YPos), color.RGBA{0, 0, 0, 0})
 			}
 		}
 	}
